Bug Found and Fixed
Found a bug in the class DefaultStringifier while retrofitting its test class TestDefaultStringifier.
When the method storeArray() is tested against a value-set passing an empty array as input, the code crashes with the ArrayIndexOutOfBounds exception.
Reason:- Line 162 of the method uses the 1st input array element to get the class for the stringifier without checking for an empty array, and hence, causes the exception to throw.

As a fix, added a check for the same.