Bug Found, Fixed and updated comment
Found a bug in the method getBaseName of class KeyProvider while retrofitting its test class TestKeyProvider.
When the method getBaseName is tested against a value-set passing a null string as input, the code crashes with an unhandled NullPointerException
Exception originates on line 649 of the method getBaseName when the code tries to find the index of the last occurrence of '@' in the input string.
As a fix, a null check for the input string was added.