Bug Found and Fixed
Found a bug in the method configure of class BenchData inside class RawErasureCoderBenchmark while retrofitting the test class TestRawErasureCoderBenchmark.
When the method configure is tested against a value-set providing -ve values for its input param, dataSizeMB, then instead of halting or giving an error, it runs successfully, giving -ve values of data and throughput in output.

In the method configure, unlike the other input chunkSizeKB
, there are no explicit checks for -ve values of dataSizeMB.

As a fix, an argument check was added for dataSizeMB, which handles -ve values by throwing an IllegalArgumentException.